Reduced the TLB (Translation lookaside buffer) from 28MB to 20MB by change 2x 1MB arrays from 64-bit longs to 32-bit ints.  I do not believe this will have any negative side effects.  Windows 3.11 and Windows 98 both still work.
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@458 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/cpu/core_dynamic/CodePageHandlerDynRec.java
@@ -221,12 +221,12 @@ final public class CodePageHandlerDynRec extends Paging.PageHandler {
 		return null;	// none found
 	}
 
-	public /*HostPt*/long GetHostReadPt(/*Bitu*/int phys_page) {
-		hostmem=(int)old_pagehandler.GetHostReadPt(phys_page);
+	public /*HostPt*/int GetHostReadPt(/*Bitu*/int phys_page) {
+		hostmem=old_pagehandler.GetHostReadPt(phys_page);
 		return hostmem;
 	}
 
-	public /*HostPt*/long GetHostWritePt(/*Bitu*/int phys_page) {
+	public /*HostPt*/int GetHostWritePt(/*Bitu*/int phys_page) {
 		return GetHostReadPt( phys_page );
 	}
